% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_binding_freqs.R
\name{calc_binding_freqs}
\alias{calc_binding_freqs}
\title{Calculate the binding frequencies}
\usage{
calc_binding_freqs(epitopeome)
}
\arguments{
\item{epitopeome}{an epitopeome}
}
\value{
a map of binding distance to fraction of binds at that
  distance
}
\description{
Calculate the binding frequencies
}
\author{
Richel J.C. Bilderbeek
}
