% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_binding_freqs_cpp}
\alias{calc_binding_freqs_cpp}
\title{Calculates the binding frequencies}
\usage{
calc_binding_freqs_cpp(epitopeome)
}
\arguments{
\item{epitopeome}{an epitopeome, without protein names}
}
\value{
a map,
  with the distance to the center of a TMH as its key
  (either x.0 or x.5),
  and the fraction of peptides at that distance binding to MHC2
}
\description{
Calculates the binding frequencies
}
\author{
Richel J.C. Bilderbeek
}
