% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_epitopeome_file_cpp}
\alias{create_epitopeome_file_cpp}
\title{Creates a FASTA file with only the TMHs}
\usage{
create_epitopeome_file_cpp(
  tmhs_filename,
  netmhc2pan_filename,
  epitopeome_filename
)
}
\arguments{
\item{tmhs_filename}{file that contains the TMHs as created by TMHMM}

\item{netmhc2pan_filename}{file that contains the NetMHC2pan results}

\item{epitopeome_filename}{name of the create epitopeome file, 
saved in FASTA format}
}
\value{
nothing
}
\description{
Creates a FASTA file with only the TMHs
}
\author{
Richel J.C. Bilderbeek
}
