% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_locatome_cpp}
\alias{create_locatome_cpp}
\title{Creates a FASTA file with only the TMHs for a TMHMM file
as created by a locally installled TMHMM binary}
\usage{
create_locatome_cpp(tmhs_filename)
}
\arguments{
\item{tmhs_filename}{file that contains the TMHs
as created by a locally installed TMHMM binary}
}
\value{
a dictionary with sequence names as key,
  and a 'locatome' as values.
}
\description{
Creates a FASTA file with only the TMHs for a TMHMM file
as created by a locally installled TMHMM binary
}
\author{
Richel J.C. Bilderbeek
}
