% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{split_str}
\alias{split_str}
\title{Split a string}
\usage{
split_str(input, seperator)
}
\arguments{
\item{input}{the input string}

\item{seperator}{the seperator}
}
\value{
a vector of strings
}
\description{
Split a string
}
\author{
Richel J.C. Bilderbeek
}
