% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epitopeome.R
\name{create_epitopeome}
\alias{create_epitopeome}
\title{Create epitopeome indicating strong MHC2 binders}
\usage{
create_epitopeome(fasta_filename, alleles = "DRB1_0101",
  tmhmm_filename = tempfile(fileext = "txt"),
  netmhc2pan_filename = tempfile(fileext = ".csv"))
}
\description{
Create epitopeome indicating strong MHC2 binders
}
