% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_epitopeome.R
\name{create_epitopeome}
\alias{create_epitopeome}
\title{Create epitopeome indicating strong MHC2 binders}
\usage{
create_epitopeome(fasta_filename, alleles = "DRB1_0101",
  tmhmm_filename = tempfile(fileext = "txt"),
  netmhc2pan_filename = tempfile(fileext = ".csv"))
}
\arguments{
\item{fasta_filename}{the name of a FASTA file with protein sequences}

\item{alleles}{one or more alleles, e.g. \code{DRB1_0101}.
See \link{get_netmhc2pan_alleles} for a full list}

\item{tmhmm_filename}{name of the file to store the
TMHMM output.
By default, this filename is generated by \link{tempfile},
and will be put in the \code{/tmp} folder,
which is cleaned by the operating system}

\item{netmhc2pan_filename}{name of the file to store the
NetMHC2pan output.
By default, this filename is generated by \link{tempfile},
and will be put in the \code{/tmp} folder,
which is cleaned by the operating system}
}
\description{
Create epitopeome indicating strong MHC2 binders
}
\author{
Richel J.C. Bilderbeek
}
