% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_epitopeome_cpp}
\alias{create_epitopeome_cpp}
\title{Creates an 'epitopeome' from a TMHMM and NetMHC2pan file}
\usage{
create_epitopeome_cpp(tmhs_filename, netmhc2pan_filename)
}
\arguments{
\item{tmhs_filename}{file that contains the TMHs as created by TMHMM}

\item{netmhc2pan_filename}{file that contains the NetMHC2pan results}
}
\value{
a map, with the sequence name as key, and its 'epitopeome'
  as value
}
\description{
Creates an 'epitopeome' from a TMHMM and NetMHC2pan file
}
\author{
Richel J.C. Bilderbeek
}
